requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.4
numpy==1.24.4
pandas==2.0.3
pkg_resources==0.0.0
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
